Refactor folder structure for the Home feature to improve organization and maintainability.
