+++ app/src/main/java/com/doganur/movieapp/ui/home/component/MovieItem.kt
@@ -89,3 +89,4 @@ fun MovieItem(
         }
     }
 }
+
